Can be compiled with OMK
All configuration that was previously done with preprocessor symbols
is now done with a new data structure macan_config.

Additionally, long-term key was also moved to the configuration structure
in favor of removed macan_set_ltk() function.

It was necessary to comment out challenge sending code in key server,
because it cannot work now.