Timer server requires configuration compiled as a shared library
In order to print meaningful debug information (packet dumps) and
implement packet filtering reasonably, it is necessary to have full
MaCAN configuration at the time server.

The simplest solution I could come up with is to compile the macan_config
structure to a shared library and dlopen() is from the key server.