Add decoding of REQ_CHALLENGE message
The format of the message seems not to be described in the paper, so this
is taken from VW implementation.

However, it seems that we generate this message in a wrong way
(fwd_id is nonsense and we probably add MAC). Actually, I'm not sure why
our node generate this message at all. Only key server should generate
this!