Add socket file descriptor to macan_ctx
This simplifies the code, because the "s" parameter was needed almost
everywhere. Because the macan_ctx needs to be passed everywhere as well,
this change reduces the number of parameters of most functions by one.

This change will also be handy during planned conversion to an event loop
library. It will be sufficient to pass just one parameter (ctx) with every
event, instead of two.