Convert the code to use libev (event loop library)
The goal here is to be able to run multiple nodes in the same
(single-threaded) process to make testing and debugging easier.

Linux version seems to be fully converted and 4signals demo works. TriCore
version is currently broken. The plan is to implement macan_ev_* functions
for TriCore to provide basic functionality of libev.

Now, most of the time and key server code is in libmacan. The binaries
keyserver and timeserver are only responsible for command line parsing and
initializing macan based on the parameters.