Initialize MaCAN time before starting the event loop
Without this change, the first time frame was send at time 0 and the
second at 0.999. This caused the value 0 to be sent two times in a row
(because floor division is used).

Rationale: In general the time server's notion of time (MaCAN time) is
synchronized with event loop timing, but the origin (time 0) is not the
same. If MaCAN time starts later than event loop time, then due to
possible jitter in callback invocation, the time broadcast callback may
be invoked just before or slightly after the increment of MaCAN time.
By starting the MaCAN time before the event loop is entered, the event
loop guarantees that the callback is invoked always after the MaCAN time
increment.

Beware that the above guarantee may be violated, if the broadcast period
is too short or of the event loop is started about the whole number of
periods after time server initialization.