Add skey_callback and improve time initialization
To establish authenticated communication faster (and to not see so many
warnings in logs), it is useful to perform some action just when a session
key is received.

Time server uses this to answer requests for authenticated time that were
sent before the key was ready. For that, the time server has to remember
the sent challenges in ctx->ts.auth_req.

Nodes use this for requesting authenticated time just after the time server
session key has been delivered to them.