Move generic source for stm32 out of the macan directory
Basically, this means that all 3rd party source code (libnettle code
from ST for STM32) is moved to the root directory. Macan directory
contains only MaCAN-specific code. This simplifies the Makefile.omk of
the macan library.

TODO: Applications (demos) will need to be modified to link not only the
macan library but also the new libraries - nettle and stm32rt.