stm32: Move system stub to nosys library
This solves problems with missing symbols like _exit. When nosys.spec is
mentioned on gcc command line, libnosys is automatically linked with the
program the right way (i.e. in group with -lc).

Also, due to this, stm32rt library was renamed to stm32, because it no
longer contains "run time" support.